* Move StatusIndicator to notebook_utils.py. * Make region configurable by text field, in all notebooks. * Stop using setup_forecast_permissions.py, use util.get_or_create_role_arn() instead. * Use conda_python3 as kernel (to prevent error message on SageMaker notebook environment.)
