+++ notebooks/util/notebook_utils.py
@@ -1,11 +1,9 @@
 import sys
 import io
-
 import ipywidgets
 
 
 widget_table = {}
-
 
 def create_text_widget( name, placeholder, default_value="" ):
 
@@ -17,4 +15,27 @@ def create_text_widget( name, placeholder, default_value="" ):
     display(widget)
     
     return widget
+
+
+class StatusIndicator:
+    
+    def __init__(self):
+        self.previous_status = None
+        self.need_newline = False
+        
+    def update( self, status ):
+        if self.previous_status != status:
+            if self.need_newline:
+                sys.stdout.write("\n")
+            sys.stdout.write( status + " ")
+            self.need_newline = True
+            self.previous_status = status
+        else:
+            sys.stdout.write(".")
+            self.need_newline = True
+        sys.stdout.flush()
+
+    def end(self):
+        if self.need_newline:
+            sys.stdout.write("\n")
